+++ CHANGELOG.md
@@ -4,8 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - SNAPSHOT
+### Added
+- Schema validation support via `SchemaManager` interface with default Confluent implementation provided (#41)
+
+## [0.3.2] - 20181216
+### Changed
+- Updated README to something that outlines this a bit better. (#54) 
+- Changed .travis.yml and added a setup script to securely send credentials for sigining and deploying master builds. (#57)
+- Added appropriate variables so that all encrypted keys work. (#57)
+
 ## [0.3.1] - 20181212
-## Changed
+### Changed
 - Specified `deploy` goal in maven-release-plugin
 - Fixed "site breaking release process" (#47)
 - Fixed ossrh release requirements  (#50)
@@ -20,5 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Shell script for build commands (#43)
 - Topic Config update functionality
 
+[0.4.0]: https://github.com/HomeAway/stream-registry/compare/v0.3.2...v0.4.0
+[0.3.2]: https://github.com/HomeAway/stream-registry/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/HomeAway/stream-registry/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/HomeAway/stream-registry/compare/v0.2.42...v0.3.0